requirements.txt
requests==2.31.0
alpaca-trade-api==3.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
beautifulsoup4==4.12.2
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.18

